refactor: lazy load editor completion focus commands, move focus commands to virtual list module
