+++ packages/renderer-worker/src/parts/ViewletEditorCompletion/ViewletEditorCompletion.ipc.js
@@ -1,15 +1,21 @@
 import * as ViewletEditorCompletion from './ViewletEditorCompletion.js'
 
+export const name = 'EditorCompletion'
+
 // prettier-ignore
 export const Commands = {
   // 'EditorCompletion.selectCurrent': Viewlet.wrapViewletCommand('EditorCompletion', ViewletEditorCompletion.selectCurrent),
   dispose:  ViewletEditorCompletion.dispose,
-  focusFirst: ViewletEditorCompletion.focusFirst,
-  focusLast:  ViewletEditorCompletion.focusLast,
-  focusNext: ViewletEditorCompletion.focusNext,
-  focusPrevious:  ViewletEditorCompletion.focusPrevious,
-  handleWheel: ViewletEditorCompletion.handleWheel,
-  selectIndex:  ViewletEditorCompletion.selectIndex,
+}
+
+export const LazyCommands = {
+  focusIndex: () => import('../VirtualList/VirtualListFocusIndex.js'),
+  focusFirst: () => import('../VirtualList/VirtualListFocusFirst.js'),
+  focusLast: () => import('../VirtualList/VirtualListFocusLast.js'),
+  focusNext: () => import('../VirtualList/VirtualListFocusNext.js'),
+  focusPrevious: () => import('../VirtualList/VirtualListFocusPrevious.js'),
+  handleWheel: () => import('../VirtualList/VirtualList.js'),
+  selectIndex: () => import('./ViewletEditorCompletionSelectIndex.js'),
 }
 
 export const Css = '/css/parts/ViewletEditorCompletion.css'
