+++ packages/renderer-worker/src/parts/VirtualList/VirtualListFocusFirst.js
@@ -1,0 +1,5 @@
+import { focusIndex } from './VirtualListFocusIndex.js'
+
+export const focusFirst = (state) => {
+  return focusIndex(state, 0)
+}
