+++ packages/renderer-worker/src/parts/VirtualList/VirtualListFocusIndex.js
@@ -1,0 +1,6 @@
+export const focusIndex = (state, index) => {
+  return {
+    ...state,
+    focusedIndex: index,
+  }
+}
