+++ packages/renderer-worker/src/parts/VirtualList/VirtualListFocusLast.js
@@ -1,0 +1,6 @@
+import { focusIndex } from './VirtualListFocusIndex.js'
+
+export const focusLast = (state) => {
+  const { items } = state
+  return focusIndex(state, items.length - 1)
+}
