+++ packages/renderer-worker/src/parts/VirtualList/VirtualListFocusPrevious.js
@@ -1,0 +1,10 @@
+import { focusIndex } from './VirtualListFocusIndex.js'
+import { focusLast } from './VirtualListFocusLast.js'
+
+export const focusPrevious = (state) => {
+  const { focusedIndex } = state
+  if (focusedIndex === 0) {
+    return focusLast(state)
+  }
+  return focusIndex(state, focusedIndex - 1)
+}
