+++ packages/renderer-worker/test/VirtualListFocusFirst.test.js
@@ -1,0 +1,11 @@
+import * as VirtualListFocusFirst from '../src/parts/VirtualList/VirtualListFocusFirst.js'
+
+test('focusFirst', () => {
+  const state = {
+    items: [1, 2, 3],
+    focusedIndex: 2,
+  }
+  expect(VirtualListFocusFirst.focusFirst(state)).toMatchObject({
+    focusedIndex: 0,
+  })
+})
